Add and move files
