+++ DNDCore/src/main/java/com/ericarao/dnd/core/model/NetworkPacket.java
@@ -1,0 +1,4 @@
+package com.ericarao.dnd.core.model;
+
+public interface NetworkPacket {
+}
