+++ DNDCore/src/main/java/com/ericarao/dnd/core/model/PacketType.java
@@ -1,0 +1,4 @@
+package com.ericarao.dnd.core.model;
+
+public enum PacketType {
+}
